refactor: move slider.js functions into vue file
the functions no longer need to be shared since refactoring RangeSlider, so there's no point in having an extra file
